+++ CHANGELOG.md
@@ -9,6 +9,7 @@ You can also check [on GitHub](https://github.com/nextcloud/news/releases), the 
 ### Changed
 
 ### Fixed
+- only use Last-Modified header if a feed supports it
 
 # Releases
 ## [26.0.2] - 2025-06-29